Return and Validation Changes

+++ js/approvals.js
@@ -17,6 +17,17 @@ function loadRequestApprovals() {
         console.log("Sorted requestedBooks : ", requestedBooks);
         for (var itr = 0; itr < requestedBooks.length; itr++) {
             let sbook = requestedBooks[itr];
+            var action = '';
+            var action2 = '';
+            var status = '';
+            if (sbook.status == "RETURNED") {
+                status = "Returned";
+                action = '<input type="button" value="Accept" onclick="acceptReturn(' + sbook.requestId + ');">';
+            } else {
+                status = "Requested";
+                action = '<input type="button" value="Approve" onclick="approve(' + sbook.requestId + ',\'A\');">';
+                action2 = '<input type="button" value="Reject" onclick="approve(' + sbook.requestId + ', \'R\');"></input>';
+            }
             row += '        <tr class="row">'
                 + '            <td class="col-xs-6 col-md-3"><b>Request Id</b></td>'
                 + '            <td class="col-xs-6 col-md-3">' + sbook.requestId + '</td>'
@@ -33,13 +44,13 @@ function loadRequestApprovals() {
                 + '            <td class="col-xs-6 col-md-3"><b>Requested By</b></td>'
                 + '            <td class="col-xs-6 col-md-3">' + sbook.requestedBy + '</td>'
                 + '            <td class="col-xs-6 col-md-3"><b>Status</b></td>'
-                + '            <td class="col-xs-6 col-md-3">' + sbook.status + '</td>'
+                + '            <td class="col-xs-6 col-md-3">' + status + '</td>'
                 + '        </tr>'
                 + '        <tr class="row">'
                 + '            <td class="col-xs-6 col-md-3"><b>Actions</b></td>'
-                + '            <td class="col-xs-6 col-md-3"><input type="button" value="Approve" onclick="approve(' + sbook.requestId + ',\'A\');"></td>'
+                + '            <td class="col-xs-6 col-md-3">' + action + '</td>'
                 + '            <td class="col-xs-6 col-md-3"></td>'
-                + '            <td class="col-xs-6 col-md-3"><input type="button" value="Reject" onclick="approve(' + sbook.requestId + ', \'R\');"></td>'
+                + '            <td class="col-xs-6 col-md-3">' + action2 + '</td>'
                 + '        </tr>';
         }
         table += row;
@@ -79,6 +90,24 @@ function approve(requestId, type) {
             selectedRequests[0].status = "ACANCELLED";
         }
         alert(messageText);
+        requestedBooks = JSON.stringify(requestedBooks);
+        console.log("requestedBooks :", requestedBooks);
+        localStorage.setItem("requestedBooks", requestedBooks);
+        loadRequestApprovals();
+    }
+}
+function acceptReturn(requestId) {
+    console.log("accept callaed");
+    let requestedBooks = localStorage.getItem("requestedBooks");
+    requestedBooks = JSON.parse(requestedBooks);
+    console.log("requestedBooks before filter : ", requestedBooks);
+    var confirmationmessege = confirm("Are you sure, you want to accept the request " + requestId + "?");
+    if (confirmationmessege) {
+        requestedBooks = requestedBooks.map(bk => {
+            if (bk.requestId == requestId) bk.status = "ACCEPTED";
+            return bk;
+        });
+        console.log("requestedBooks :", requestedBooks);
         requestedBooks = JSON.stringify(requestedBooks);
         console.log("requestedBooks :", requestedBooks);
         localStorage.setItem("requestedBooks", requestedBooks);
